+++ packages/ui/src/components/data-editor/editor-items/index.js
@@ -3,4 +3,3 @@ export * from "./TitleItem";
 export * from "./GridItem";
 export * from "./ButtonItem";
 export * from "./ColorItem";
-export * from "./ToggleButtonItem";